Refactor: relocate example feature
